Update unit tests using binary to integration tests
