+++ Cargo.toml
@@ -10,11 +10,11 @@ license = "MIT"
 categories = ["command-line-utilities", "text-processing"]
 
 [dependencies]
-anstyle = "1.0.7"
-clap = { version = "4.5", features = ["derive", "unicode"] }
-human-panic = "2.0"
-rand = "0.8"
+anstyle = "1.0.14"
+clap = { version = "4.6.0", features = ["derive", "unicode"] }
+human-panic = "2.0.6"
+rand = "0.10.0"
 
 [dev-dependencies]
-assert_cmd = "2.0"
-predicates = "3.1"
+assert_cmd = "2.2.0"
+predicates = "3.1.4"
